Add support for installed-tests metadata files
The installed-tests standard requires .test files to be generated for
each installed test. We don’t actually know the variable which lists all
the installed tests, but almost all modules use TESTS to list them too.

https://wiki.gnome.org/Initiatives/GnomeGoals/InstalledTests

This could cause problems if a test is listed in TESTS but a similar one
with a .test extension also exists (e.g. if TEST_EXTENSIONS is being
used). That is a bit of a corner case though.