Avoid errors if $(LTCOMPILE) has shell code in it
In https://github.com/ostreedev/ostree we have some shell quoting
in $(AM_CPPFLAGS) like `-DOSTREE_FEATURES='"$(OSTREE_FEATURES)"'`.
This ends up in `$(LTCOMPILE)` which causes shell errors since we
are trying to quote outside of it.

Fix this by using `findstring` to pick out the string libtool if it
exists, which is obviously shell-safe.